some of the cleanups are in src  not the root of the checkout
add src/ to the rm's for the panama5 pidp10-test scansw10 sty33
as that is where they are. though realistically we should just
leave everything there as it will all be added back when doing a
"git pull --rebase" Ideally we had a git repo with just the sources for
the PiDP-10 without everything else.

Signed-off-by: Dennis Gilmore <dennis@ausil.us>